Some minor changes - see description
Changed delay time of laser to 0.20 to make it consistent with what is in the inspector. Increased speed of asteroid from 1 to 3.  Made collision box for asteroids more accurate. The solution to the collision issue seems to be a lower velocity for the ship laser, and 10 appears to be the ideal number for this.