Enabled settings button to work with both mouse click/touch
Removed the old settings button from the scene - that is the change to Level1.unity. Replaced it with a settings button that works with both mouse click/touch. Fixed timestep needs to be at 0.001, not 0.0001, or the game gets too laggy. Not sure why it was at 0.0001, but I changed it back to 0.001.